pyproject.toml
[project]
name = "asset-auto-picker"
readme = "README.md"
version = "1.2.0"
description = "The MCP Server use for auto pick file"
requires-python = ">=3.11"
authors = [
  { name = "STrong" },
]
license = { text = "MIT" }
dependencies = [
    "colorama>=0.4.6",
    "fastmcp>=2.13.0.2",
    "fore>=0.1.8",
    "httpx>=0.28.1",
    "mcp>=1.19.0",
    "rapidfuzz>=3.14.1",
    "numpy>=1.26.0",
    "sentence-transformers>=2.2.2",
    "qdrant-client>=1.10.0",
    "python-dotenv>=1.0.1",
    "openai>=1.40.0",
    "twine>=6.2.0",
    "openai-agents==0.4.2",
]
keywords = [
  "mcp",
  "model-context-protocol",
  "mcp-name: io.github.TranThienTrong/asset-auto-picker",
]
